Enable -Wformat=2 and fix format-string warnings
Add -Wall -Wformat=2 (with -Wno-format-nonliteral / -truncation /
-invalid-offsetof to silence known-benign cases) so format-string
mismatches fail the build instead of silently misprinting.

Mechanical fixes across the slscore/ tree:
  - %d -> %zu for size_t (.size(), strlen())
  - %d -> %lu for pthread_t (with explicit cast)
  - %d -> %lld for int64_t (with explicit cast)
  - drop dead `if (NULL == arr)` checks on stack arrays
  - add missing `this` argument in SLSRelayManager::connect format
  - virtual destructors on CSLSThread / CSLSEpollThread / CSLSGroup
  - sls_log gets __attribute__((format(printf,2,3))) so future
    drift fails at compile time

One real bug along the way: `std:list <CSLSManager*>` (single colon,
parsed as a label) -> `std::list<CSLSManager*>` in srt-live-server.cpp.

No behavior change.

+++ slscore/SLSEpollThread.cpp
@@ -53,7 +53,7 @@ int CSLSEpollThread::init_epoll()
 
     m_eid = CSLSSrt::libsrt_epoll_create();
     if (m_eid < 0) {
-        sls_log(SLS_LOG_INFO, "[%p]CSLSEpollThread::work, srt_epoll_create failed. th_id=%lld.", this, m_th_id);
+        sls_log(SLS_LOG_INFO, "[%p]CSLSEpollThread::work, srt_epoll_create failed. th_id=%lu.", this, (unsigned long)m_th_id);
         return CSLSSrt::libsrt_neterrno();
     }
     //compatible with srt v1.4.0 when container is empty.
@@ -66,7 +66,7 @@ int CSLSEpollThread::uninit_epoll()
     int ret = 0;
     if (m_eid >= 0) {
         CSLSSrt::libsrt_epoll_release(m_eid);
-        sls_log(SLS_LOG_INFO, "[%p]CSLSEpollThread::work, srt_epoll_release ok, m_th_id=%lld.", this, m_th_id);
+        sls_log(SLS_LOG_INFO, "[%p]CSLSEpollThread::work, srt_epoll_release ok, m_th_id=%lu.", this, (unsigned long)m_th_id);
     }
     return ret;
 }
@@ -74,14 +74,14 @@ int CSLSEpollThread::uninit_epoll()
 int CSLSEpollThread::work()
 {
     int ret = 0;
-    sls_log(SLS_LOG_INFO, "[%p]CSLSEpollThread::work, begin th_id=%lld.", this, m_th_id);
+    sls_log(SLS_LOG_INFO, "[%p]CSLSEpollThread::work, begin th_id=%lu.", this, (unsigned long)m_th_id);
 	//epoll loop
 	while (!m_exit) {
 	    handler();
 	}
 
 	clear();
-    sls_log(SLS_LOG_INFO, "[%p]CSLSEpollThread::work, end th_id=%lld.", this, m_th_id);
+    sls_log(SLS_LOG_INFO, "[%p]CSLSEpollThread::work, end th_id=%lu.", this, (unsigned long)m_th_id);
 	return ret;
 }
